FINISH BONUS

+++ srcs_n/bitshift.c
@@ -6,13 +6,13 @@
 /*   By: jcourtoi <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/27 15:24:09 by jcourtoi          #+#    #+#             */
-/*   Updated: 2022/05/27 15:24:59 by jcourtoi         ###   ########.fr       */
+/*   Updated: 2022/05/27 16:02:40 by jcourtoi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes_n/so_long.h"
 
-int	create_trgb(int t, int r, int g, int b)
+int	create_trgb(long int t, int r, int g, int b)
 {
 	return (t << 24 | r << 16 | g << 8 | b);
 }
